+++ web/js/annotation_sidebar/AnnotationSidebar.tsx
@@ -207,7 +207,7 @@ export class AnnotationSidebar extends React.Component<IProps, IState> {
 
     private onExport(path: string, format: ExportFormat) {
 
-        Exporters.doExport(path, format, this.props.doc.docMeta)
+        Exporters.doExport(path, this.props.persistenceLayerProvider(), format, this.props.doc.docMeta)
             .catch(err => log.error(err));
 
     }